+++ connectivity/FEATURE_BLE/include/ble/SecurityManager.h
@@ -27,8 +27,6 @@
 #include "ble/common/ble/blecommon.h"
 
 #include "ble/common/ble/BLETypes.h"
-#include "ble/internal/PalConnectionMonitor.h"
-#include "ble/internal/PalSecurityManager.h"
 
 namespace ble {
