+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/include/ble/internal/BLEInstanceBaseImpl.h
@@ -25,16 +25,16 @@
 
 #include "CordioHCIDriver.h"
 #include "ble/GattServer.h"
-#include "PalAttClient.h"
-#include "PalGattClient.h"
+#include "source/pal/PalAttClient.h"
+#include "source/pal/PalGattClient.h"
 #include "ble/GattClient.h"
-#include "ble/internal/PalGap.h"
+#include "source/pal/PalGap.h"
 #include "ble/Gap.h"
-#include "PalGenericAccessService.h"
+#include "source/pal/PalGenericAccessService.h"
 #include "ble/SecurityManager.h"
-#include "PalEventQueue.h"
+#include "source/pal/PalEventQueue.h"
 #include "drivers/LowPowerTimer.h"
-#include "ble/internal/PalSecurityManager.h"
+#include "source/pal/PalSecurityManager.h"
 
 #include "source/generic/GapImpl.h"
 #include "source/generic/GattClientImpl.h"
