+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/include/ble/internal/PalSigningMonitorImpl.h
@@ -19,7 +19,7 @@
 #ifndef IMPL_PAL_SIGNING_MONITOR_H_
 #define IMPL_PAL_SIGNING_MONITOR_H_
 
-#include "ble/internal/PalSigningMonitor.h"
+#include "source/pal/PalSigningMonitor.h"
 
 namespace ble {
