+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/source/BLEInstanceBaseImpl.cpp
@@ -37,10 +37,10 @@
 #include "mbed_assert.h"
 #include "bstream.h"
 
-#include "ble/internal/PalAttClient.h"
-#include "ble/internal/PalSecurityManager.h"
-#include "ble/internal/PalGap.h"
-#include "ble/internal/PalSigningMonitor.h"
+#include "source/pal/PalAttClient.h"
+#include "source/pal/PalSecurityManager.h"
+#include "source/pal/PalGap.h"
+#include "source/pal/PalSigningMonitor.h"
 #include "source/BLEInstanceBase.h"
 #include "CordioHCIDriver.h"
 #include "source/generic/GattServerImpl.h"
