+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/source/PalEventQueueImpl.cpp
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-#include "ble/internal/PalEventQueue.h"
+#include "source/pal/PalEventQueue.h"
 #include "source/BLEInstanceBase.h"
 
 namespace ble {
