+++ connectivity/FEATURE_BLE/source/generic/SecurityManagerImpl.h
@@ -30,9 +30,9 @@
 #include "ble/common/ble/GapTypes.h"
 #include "ble/common/ble/BLETypes.h"
 #include "SecurityDb.h"
-#include "ble/internal/PalConnectionMonitor.h"
-#include "ble/internal/PalSigningMonitor.h"
-#include "ble/internal/PalSecurityManager.h"
+#include "source/pal/PalConnectionMonitor.h"
+#include "source/pal/PalSigningMonitor.h"
+#include "source/pal/PalSecurityManager.h"
 #include "ble/SecurityManager.h"
 
 namespace ble {
