+++ connectivity/FEATURE_BLE/source/pal/PalEventQueue.h
