+++ connectivity/FEATURE_BLE/source/pal/PalGap.h
