+++ connectivity/FEATURE_BLE/source/pal/PalSecurityManager.h
