+++ connectivity/FEATURE_BLE/source/pal/PalSigningMonitor.h
